% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{saxoapi-package}
\alias{saxoapi}
\alias{saxoapi-package}
\title{saxoapi: Interact with Saxo API}
\description{
Interact with Saxo Bank Open API. Make Oauth 2.0 authentification, wrap most important API endpoints (account info, orders etc.).
}
