% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaxoAuthService.R
\name{SaxoAuthService}
\alias{SaxoAuthService}
\title{SaxoAuthService Class}
\description{
OPerates with Saxo Open API.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{app_config}}{Saxo APP object}

\item{\code{auth_redirect_url}}{Redirect URL for Saxo API.}

\item{\code{auth_code}}{Authentification code}

\item{\code{token_data}}{Token data}

\item{\code{base_request}}{Base httr2 request object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SaxoAuthService-new}{\code{SaxoAuthService$new()}}
\item \href{#method-SaxoAuthService-login}{\code{SaxoAuthService$login()}}
\item \href{#method-SaxoAuthService-my_oauth_flow_auth_code_listen}{\code{SaxoAuthService$my_oauth_flow_auth_code_listen()}}
\item \href{#method-SaxoAuthService-exercise_authorization}{\code{SaxoAuthService$exercise_authorization()}}
\item \href{#method-SaxoAuthService-get_refreshed_token}{\code{SaxoAuthService$get_refreshed_token()}}
\item \href{#method-SaxoAuthService-get_base_request}{\code{SaxoAuthService$get_base_request()}}
\item \href{#method-SaxoAuthService-request_get_saxo}{\code{SaxoAuthService$request_get_saxo()}}
\item \href{#method-SaxoAuthService-get_user}{\code{SaxoAuthService$get_user()}}
\item \href{#method-SaxoAuthService-get_client}{\code{SaxoAuthService$get_client()}}
\item \href{#method-SaxoAuthService-get_accounts}{\code{SaxoAuthService$get_accounts()}}
\item \href{#method-SaxoAuthService-get_balances_user}{\code{SaxoAuthService$get_balances_user()}}
\item \href{#method-SaxoAuthService-get_balance_account}{\code{SaxoAuthService$get_balance_account()}}
\item \href{#method-SaxoAuthService-get_instruments}{\code{SaxoAuthService$get_instruments()}}
\item \href{#method-SaxoAuthService-place_order}{\code{SaxoAuthService$place_order()}}
\item \href{#method-SaxoAuthService-get_orders_all}{\code{SaxoAuthService$get_orders_all()}}
\item \href{#method-SaxoAuthService-get_positions}{\code{SaxoAuthService$get_positions()}}
\item \href{#method-SaxoAuthService-get_net_positions_user}{\code{SaxoAuthService$get_net_positions_user()}}
\item \href{#method-SaxoAuthService-clone}{\code{SaxoAuthService$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaxoAuthService-new"></a>}}
\if{latex}{\out{\hypertarget{method-SaxoAuthService-new}{}}}
\subsection{Method \code{new()}}{
Create a new AuthService object with provided AppConfig.
When initialized, config is loaded either directly from app_config
argument or from "app_config.json".
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaxoAuthService$new(
  app_config = NULL,
  auth_redirect_url = NULL,
  auth_code = NULL,
  token_data = NULL,
  base_request = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{app_config}}{Saxo APP object.}

\item{\code{auth_redirect_url}}{Redirect URL for Saxo API.}

\item{\code{auth_code}}{Authentification code.}

\item{\code{token_data}}{Token data.}

\item{\code{base_request}}{Base httr2 request.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `SaxoAuthService` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaxoAuthService-login"></a>}}
\if{latex}{\out{\hypertarget{method-SaxoAuthService-login}{}}}
\subsection{Method \code{login()}}{
Login to Sax Open API
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaxoAuthService$login(
  redirect_url = NULL,
  redirect_port = NULL,
  user_name = NULL,
  password = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{redirect_url}}{Redirect url}

\item{\code{redirect_port}}{Redirect port}

\item{\code{user_name}}{User name for authentification in the browser.
NULL if you want to authenticate manually.}

\item{\code{password}}{Password for authentification in the browser.
NULL if you want to authenticate manually.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Object that can get token.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaxoAuthService-my_oauth_flow_auth_code_listen"></a>}}
\if{latex}{\out{\hypertarget{method-SaxoAuthService-my_oauth_flow_auth_code_listen}{}}}
\subsection{Method \code{my_oauth_flow_auth_code_listen()}}{
Starts an webserver that listens for the response from the resource server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaxoAuthService$my_oauth_flow_auth_code_listen(
  host_ip = "127.0.0.1",
  port = 4321,
  path = "/redirect"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{host_ip}}{IP address web server will be bound to.}

\item{\code{port}}{Port to bind web server to.
By default, this uses a random port.
You may need to set it to a fixed port if the API requires that
the redirect_uri specified in the client exactly matches
the redirect_uri generated by this function.}

\item{\code{path}}{Add path to listener}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Object that can get token.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaxoAuthService-exercise_authorization"></a>}}
\if{latex}{\out{\hypertarget{method-SaxoAuthService-exercise_authorization}{}}}
\subsection{Method \code{exercise_authorization()}}{
Get token.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaxoAuthService$exercise_authorization(auth_code)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{auth_code}}{Authorization code.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Saxo Open API token.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaxoAuthService-get_refreshed_token"></a>}}
\if{latex}{\out{\hypertarget{method-SaxoAuthService-get_refreshed_token}{}}}
\subsection{Method \code{get_refreshed_token()}}{
Refresh token endpoint https://www.developer.saxo/openapi/learn/oauth-authorization-code-grant
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaxoAuthService$get_refreshed_token(refresh_token)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{refresh_token}}{Refresh token.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Saxo Open API token.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaxoAuthService-get_base_request"></a>}}
\if{latex}{\out{\hypertarget{method-SaxoAuthService-get_base_request}{}}}
\subsection{Method \code{get_base_request()}}{
Check token and prepare request
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaxoAuthService$get_base_request()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tokens}}{Saxo Open API tokens.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Prepared httr2 request object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaxoAuthService-request_get_saxo"></a>}}
\if{latex}{\out{\hypertarget{method-SaxoAuthService-request_get_saxo}{}}}
\subsection{Method \code{request_get_saxo()}}{
Help function to send get requests
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaxoAuthService$request_get_saxo(add_url)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{add_url}}{Add url to base url.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
response from request.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaxoAuthService-get_user"></a>}}
\if{latex}{\out{\hypertarget{method-SaxoAuthService-get_user}{}}}
\subsection{Method \code{get_user()}}{
User endpoint <https://www.developer.saxo/openapi/referencedocs/root/v1/user>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaxoAuthService$get_user()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
User data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaxoAuthService-get_client"></a>}}
\if{latex}{\out{\hypertarget{method-SaxoAuthService-get_client}{}}}
\subsection{Method \code{get_client()}}{
Client Details endpoint <https://www.developer.saxo/openapi/referencedocs/port/v1/clients>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaxoAuthService$get_client()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Client Details data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaxoAuthService-get_accounts"></a>}}
\if{latex}{\out{\hypertarget{method-SaxoAuthService-get_accounts}{}}}
\subsection{Method \code{get_accounts()}}{
Accounts endpoint <https://www.developer.saxo/openapi/referencedocs/port/v1/accounts>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaxoAuthService$get_accounts()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Accounts data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaxoAuthService-get_balances_user"></a>}}
\if{latex}{\out{\hypertarget{method-SaxoAuthService-get_balances_user}{}}}
\subsection{Method \code{get_balances_user()}}{
Get balance data for logged-in client
   More info on <https://www.developer.saxo/openapi/referencedocs/port/v1/balances>.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaxoAuthService$get_balances_user()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Positions data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaxoAuthService-get_balance_account"></a>}}
\if{latex}{\out{\hypertarget{method-SaxoAuthService-get_balance_account}{}}}
\subsection{Method \code{get_balance_account()}}{
Account balance endpoint <https://www.developer.saxo/openapi/referencedocs/port/v1/balances>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaxoAuthService$get_balance_account(account_key, client_key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{account_key}}{Account key.}

\item{\code{client_key}}{client key.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Account Balance data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaxoAuthService-get_instruments"></a>}}
\if{latex}{\out{\hypertarget{method-SaxoAuthService-get_instruments}{}}}
\subsection{Method \code{get_instruments()}}{
Instruments endpoint <https://www.developer.saxo/openapi/referencedocs/ref/v1/instruments>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaxoAuthService$get_instruments(
  keywords = NULL,
  asset_types = NULL,
  include_nontradable = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{keywords}}{Keywords for instrument.}

\item{\code{asset_types}}{Asset types.}

\item{\code{include_nontradable}}{Include nontradable instrument or not.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Instruments Data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaxoAuthService-place_order"></a>}}
\if{latex}{\out{\hypertarget{method-SaxoAuthService-place_order}{}}}
\subsection{Method \code{place_order()}}{
Place New Order endpoint <https://www.developer.saxo/openapi/referencedocs/trade/v2/orders>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaxoAuthService$place_order(body_params = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{body_params}}{Body of POST request with request parameteres (list)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Order ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaxoAuthService-get_orders_all"></a>}}
\if{latex}{\out{\hypertarget{method-SaxoAuthService-get_orders_all}{}}}
\subsection{Method \code{get_orders_all()}}{
All Open Orders endpoint endpoint <https://www.developer.saxo/openapi/referencedocs/port/v1/orders>
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaxoAuthService$get_orders_all(field_groups = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{field_groups}}{FieldGroups, list that can contain:
DisplayAndFormat - Adds information about the instrument,
  which is useful for display and formatting.
  This includes Currency Code, Decimals, Instrument Description,
  Display Decimals, Price format and Symbol
ExchangeInfo - Adds information about the instrument's exchange.
  This includes Exchange name, exchange code and open status.
Greeks - Greeks for Option(s), only applicable to Fx Options,
  Contract Options and Contract options CFD}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Json object of all orders.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaxoAuthService-get_positions"></a>}}
\if{latex}{\out{\hypertarget{method-SaxoAuthService-get_positions}{}}}
\subsection{Method \code{get_positions()}}{
Get positions for a client, account group, account or a position.
   More info on <https://www.developer.saxo/openapi/referencedocs/port/v1/positions>.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaxoAuthService$get_positions(client_key = NULL, field_group = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client_key}}{client key.}

\item{\code{field_group}}{FieldGroups: DisplayAndFormat, ExchangeInfo, Greeks,
PositionBase, PositionIdOnly, PositionView.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Positions data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaxoAuthService-get_net_positions_user"></a>}}
\if{latex}{\out{\hypertarget{method-SaxoAuthService-get_net_positions_user}{}}}
\subsection{Method \code{get_net_positions_user()}}{
Get net positions for the current user's client.
   More info on <https://www.developer.saxo/openapi/referencedocs/port/v1/netpositions>.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaxoAuthService$get_net_positions_user(
  top = NULL,
  skip = NULL,
  field_group = "DisplayAndFormat"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{top}}{top.}

\item{\code{skip}}{skip.}

\item{\code{field_group}}{look at <https://www.developer.saxo/openapi/referencedocs/port/v1/netpositions>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Positions data.
#' @description Get positions for a client, account group, account or a position.
#'    More info on <https://www.developer.saxo/openapi/referencedocs/port/v1/positions>.
#'
#' @param client_key client key.
#' @param field_group FieldGroups: DisplayAndFormat, ExchangeInfo, Greeks,
#'    PositionBase, PositionIdOnly, PositionView.
#'
#' @return Positions data.
get_positions = function(client_key = NULL, field_group = NULL) {

  # make get request
  resp <- self$base_request %>%
    req_url_path_append("port/v1/positions") %>%
    req_url_query(ClientKey = client_key, FieldGroups = field_group) %>%
    req_perform()

  if (resp_status(resp) == 200) {
    resp <- resp_body_json(resp)
  }

  return(resp)
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaxoAuthService-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SaxoAuthService-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaxoAuthService$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
