% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaxoAuthService.R
\name{SaxoAuthService}
\alias{SaxoAuthService}
\title{SaxoAuthService Class}
\description{
OPerates with Saxo Open API.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{app_config}}{Sax APP object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SaxoAuthService-new}{\code{SaxoAuthService$new()}}
\item \href{#method-SaxoAuthService-login}{\code{SaxoAuthService$login()}}
\item \href{#method-SaxoAuthService-clone}{\code{SaxoAuthService$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaxoAuthService-new"></a>}}
\if{latex}{\out{\hypertarget{method-SaxoAuthService-new}{}}}
\subsection{Method \code{new()}}{
Create a new AuthService object with provided AppConfig.
When initialized, config is loaded either directly from app_config
argument or from "app_config.json".
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaxoAuthService$new(app_config = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{app_config}}{Sax APP object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `SaxoAuthService` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaxoAuthService-login"></a>}}
\if{latex}{\out{\hypertarget{method-SaxoAuthService-login}{}}}
\subsection{Method \code{login()}}{
Login to Sax Open API
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaxoAuthService$login(redirect_url = NULL, redirect_port = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{redirect_url}}{Redirect url}

\item{\code{redirect_port}}{Redirect port}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Object that can get token.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SaxoAuthService-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SaxoAuthService-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SaxoAuthService$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
